Extract an Access concern
No new code, just moving methods that generate access URLs into
their own concern.
